Move libstm32l4 files to cores/arduino
