+++ cores/arduino/libstm32l4/analog.c
