+++ cores/arduino/libstm32l4/clock.h
