+++ cores/arduino/libstm32l4/hw_config.h
