+++ cores/arduino/libstm32l4/interrupt.c
