+++ cores/arduino/libstm32l4/spi_com.h
