+++ cores/arduino/libstm32l4/stm32_eeprom.h
