+++ cores/arduino/libstm32l4/timer.c
