+++ cores/arduino/libstm32l4/uart.h
