+++ cores/arduino/libstm32l4/uart_emul.h
